CEDS-1950 Make WarehouseIdentification backwards compatible
As part of the story we have removed the `WarehouseIdentificationType`
from the user journey, but it is still required to go into the DMS
payload.

James W. and Joe will create a story to update the `hint` on the
warehouse IdentificationNumber page and to provide new validation rules
for the Warehouse type.
That ticket will be played in the following sprint.

In the meantime this change is simply not to downgrade the payload that
is sent to DMS, as DMS expects both identificationNumber and Type or
none of them.