CEDS-1955 Clean up the Warehouse Model
The work carried out by the splitting of the Warehouse page has been
done by reutilising the `WarehouseIdentification` model. This has proven
to be confusing and the `supervisingCustomsOffice` and
`inlandModeOfTransport` no longer belong to the WarehouseIdentification
model.

This change will require a change to the model in the FE application and
in the UAT/smoke/acceptance tests.

Also relates to: CEDS-1950, CEDS-1953, CEDS-1954