BAU - Understand DB Migrations on Remove Environments
The default lock, locks the collection for DB migrations for 24 hours.
If the DB migration takes longer than a 3 minutes, for some reason the
connection to the collection is lost and the lock is never removed.

By providing a quick lock, it would allow the application to try and
acquire the lock at least up to 10 times.