CEDS-2755 Add a Pointer Based URL logic
In some instances the WCO Declaration pointer will be returned but it
can map to two different pages on our user journey, which is depending on the
declaration type and the questions answered.

The approach taken is to provide a [URL_PATH] placeholder that will then
be replaced by a URL PATH resolver in the frontend.