BAU Enhance and simplify declaration submission process
After this change, Declaration document in MongoDB has
its status changed to 'complete' at the time of fetching,
instead of updating it later in the process.

This change prevents a situation where calling the
submission process twice in a short period of time
could result in double submission due to race
condition (previously, declaration status was updated
as a separate operation in Mongo, after checking the
status of fetched declaration).