[CEDS-2800] Reduce query batch size for db migrations
Currently the db migration code starts in QA but never
finishes (without throwing any errors or exceptions).

It is suspected that there might be a memory limit on
mongo on the platform that we are exceeding by trying
to run 100 update queries at a time. s

So now reducing this number to 10 to see if that makes
a difference.