[CEDS-2888] Auto set routing country to GB
DMS rules state that if the origin country is set
to GB then the routing country must also be set as
GB.

Users are understandably unaware of this rule so
have been not setting a routing country when
exporting goods from GB to other countries.

To reduce the number of declarations falling fowl
of this and being rejected we will now automatically
set the routing country to GB if the originating
country is GB and no routing countries are specified
by the user.