CEDS-3036 Implement asynchronous notifications processing
Notification parsing and then saving is now done separately from
responding to the request but it is executed in the same
ExecutionContext. Should it prove inefficient, new ExecutionContext can
be implemented at a later date.

Parsing and saving logic has been extracted from NotificationService as
this class was carrying too many responsibilities. It is now much easier
to test as well.

Being aware of incoming changes and new processes to be introduced on
the receipt of a new Notification, trait NotificationReceiptAction has
been implemented. This allows to have common interface for all these
processes.