CEDS-3045 Implement SendEmailsJob
This is a period job that sends email requests to email service.
It is based on SendEmailWorkItemRepository and SendEmailWorkItems.
Email sending process has been changed to unify treatment of email
triggers.

How email sending process looks like?

After parsing Notification, saving it into Mongo and updating related
Submission with MRN, SendEmailForDmsDocAction makes a check if this
Notification has DMSDOC status. If it has not, the process ends there.
However, if it is DMSDOC, then a new document is created in
sendEmailsWorkItems collection, containing Notification's document id
and MRN.
This process also works in a scenario where Notification arrived before
inserting Submission document into Mongo.

SendEmailsJob runs periodically with configured interval. It pulls
documents one by one from sendEmailsWorkItems that are not processed
yet, builds a request to email service, sends it and handles the result
of this call. There are 4 scenarios being handled:
1. Email Accepted - request has been accepted by email service. WorkItem
is marked as 'Succeeded'.
2. Missing Data - a piece of information required to build the request
cannot be obtained. WorkItem is marked as 'Failed'.
3. 4xx response from email service - potential problem on our side for
this particular request. WorkItem is marked as 'Failed'
4. 5xx response from email service - This usually means that there is no
reason to try sending subsequent requests. Because of that current
WorkItem is marked as 'Failed' but SendEmailsJob stops processing more
WorkItems. The next attempt will be taken with the next run.

SendEmailsJob is executed by ALL instances of this service, but using
WorkItemRepository ensures that given WorkItem is processed only
by a single instance during particular Job run.
An explanation for p. no. 4 above: during a single Job run there may be
as many requests sent as there are instances of customs-declare-exports
service before stopping the process for the rest of WorkItems.

Reasoning behind this design:

There was another proposition on the process. It assumed adding an
optional field to the existing Notification class which would contain
information necessary to handle email sending. 
However, after some trial and error it turned out that both solutions
are very similar in terms of complexity.
In the end, I decided to go this path because it allows to build
implementation with minimum interference into existing code and data
model. It also keeps email-related data separate from the actual
Notification data.
 I also wanted to avoid having all emails handled by a single instance.
In some cases this could lead to very unequal work distribution.