BAU Fix PagerDutyAlertValidatorSpec instability
For one the test cases Duration parameter was equal to
the threshold value. This was causing the test to fail
randomly.
As a solution, 1 second has been subtracted from this
Duration value. This small time difference is irrelevant
to the whole logic, but ensures the test case pass more
reliably.