BAU Improve Declaration Submission performance
It turned out that the most time consuming part of XML generation
was the creation of JAXBContent. Full problem description together
with solutions can be found here:
https://www.ibm.com/support/pages/jaxbcontext-initialization-takes-long-time

As you can read in one of the solutions, JAXBContext is bound to a
set of classes/packages and it can be re-used for repeatedly converting
the same model. This is why the context has been made a lazy field in
the class.

This upgrade improves XML generation step approx. 100 times.
Request handling time for Declaration submission is lowered
by approx. 50%.