[CEDS-3784] Optimisation bug fix
There was a bug where as the fetching of the submission record happened for each notification before any of the
conditional updates occur, this resulted in the first update being successful but all the subsequent
updates failing (due to its condition no longer being satisfied).

I've removed the condition on the update which will mean unnecessary MRN updates happening for now. Will come bakc to this for
a better solution later.