[CEDS-3854] Notification DateTime parsing bug
Fix bug when parsing dateTime values in the notification XML payload.
We were converting offset datTime values such as '20220401102910+01' into
UTC values of '2022-04-01T10:29:10Z[UTC]' (not adjusting for the offset!).

This meant that when we displayed these values to the user and adjusted for
BST we were adding +1 hour to the actual dateTime value.