CEDS-3977 Review uses of looking up Submissions by unindexed fields
Adding two new indexes, one on the submissions collection to lookup via LRN, the other is on the declarations collection to lookup via parentDeclarationId.
Also refactored the general query endpoint that used the SubmissionQueryParameters to just retrieve all submissions for a user (no additional filtering) and added a new endpoint to cover the 'is LRN used in last 48hr' use case.