CEDS-5592 Fix transformation from ExportsDeclarationRequest
The ExportsDeclarationRequest was making a field by field copy of the DeclarationMetaData object rather than just copying the entire thing. This resulted in the new `associatedSubmissionId` field being stripped out of any updates to a draft declaration.

I've also renamed the migration job that populates the new `associatedSubmissionId` field so that it runs again in production to re-populate those declarations that will now be missing the new fields.